Move the python script into a nn_wind_prediction python package

+++ wind_prediction/train.py
@@ -1,25 +1,27 @@
 #!/usr/bin/env python
 
+from __future__ import print_function
+
 import argparse
-import gc
-import models
-import numpy as np
+import nn_wind_prediction.data as data
+import nn_wind_prediction.models as models
+import nn_wind_prediction.nn as nn_custom
+import nn_wind_prediction.utils as utils
 import os
-from tensorboardX import SummaryWriter
 import time
 import torch
 import torch.nn as nn
 import torch.optim as optim
 from torch.utils.data import DataLoader
 from torch.optim.lr_scheduler import StepLR
-import utils
 
 now_time = time.strftime("%Y_%m_%d-%H_%M")
 
 parser = argparse.ArgumentParser(description='Training an EDNN for predicting wind data from terrain')
 parser.add_argument('-np', '--no-plots', dest='make_plots', action='store_false', help='Turn off plots (default False)')
 parser.add_argument('-y', '--yaml-config', required=True, help='YAML config file')
-parser.add_argument('-o', '--output-dir', default='models/trained_models/', help='Output directory')
+parser.add_argument('-o', '--output-dir', default='trained_models/', help='Output directory')
+parser.add_argument('-w', '--writer', dest='use_writer', action='store_true', help='Use a SummaryWriter to log the learningcurve')
 
 args = parser.parse_args()
 
@@ -40,9 +42,10 @@ if (os.path.isdir("/cluster/scratch/")):
     t_intermediate = time.time()
     os.system('cp '  + run_params.data['validationset_name'] + ' ' + validationset_name)
     print("INFO: Finished copying validationset in %s seconds" % (time.time() - t_intermediate))
-    t_intermediate = time.time()
-    os.system('cp '  + run_params.data['testset_name'] + ' ' + testset_name)
-    print("INFO: Finished copying testset in %s seconds" % (time.time() - t_intermediate))
+    if run_params.run['evaluate_testset']:
+        t_intermediate = time.time()
+        os.system('cp '  + run_params.data['testset_name'] + ' ' + testset_name)
+        print("INFO: Finished copying testset in %s seconds" % (time.time() - t_intermediate))
 
 else:
     print('Script is running on the a local machine')
@@ -50,15 +53,14 @@ else:
     validationset_name = run_params.data['validationset_name']
     testset_name = run_params.data['testset_name']
 
-gc.collect()
 
 # define dataset and dataloader
-trainset = utils.MyDataset(trainset_name, compressed = run_params.data['compressed'], **run_params.MyDataset_kwargs())
+trainset = data.MyDataset(trainset_name, compressed = run_params.data['compressed'], **run_params.MyDataset_kwargs())
 
 trainloader = torch.utils.data.DataLoader(trainset, batch_size=run_params.run['batchsize'],
                                           shuffle=True, num_workers=run_params.run['num_workers'])
 
-validationset = utils.MyDataset(validationset_name, compressed = run_params.data['compressed'], **run_params.MyDataset_kwargs())
+validationset = data.MyDataset(validationset_name, compressed = run_params.data['compressed'], **run_params.MyDataset_kwargs())
 
 validationloader = torch.utils.data.DataLoader(validationset, shuffle=False, batch_size=run_params.run['batchsize'],
                                           num_workers=run_params.run['num_workers'])
@@ -92,19 +94,12 @@ scheduler = StepLR(optimizer, step_size=run_params.run['learning_rate_decay_step
 
 if run_params.run['loss_function'] == 1:
     loss_fn = torch.nn.L1Loss()
-    loss_fn_val = torch.nn.L1Loss()
 elif run_params.run['loss_function'] == 2:
-    loss_fn = utils.GaussianLogLikelihoodLoss()
-    loss_fn_val = utils.GaussianLogLikelihoodLoss()
+    loss_fn = nn_custom.GaussianLogLikelihoodLoss()
 elif run_params.run['loss_function'] == 3:
     loss_fn = utils.MyLoss(device)
-    loss_fn_val = utils.MyLoss(device)
 else:
     loss_fn = torch.nn.MSELoss()
-    loss_fn_val = torch.nn.MSELoss()
-
-# initialize the tensorboard writer
-writer = SummaryWriter(os.path.join(model_dir, 'learningcurve'))
 
 # save the model parameter in the beginning
 run_params.save(model_dir)
@@ -115,99 +110,16 @@ print(' ')
 run_params.print()
 print('-----------------------------------------------------------------------------')
 
-# start the training for n_epochs
-start_time = time.time()
-for epoch in range(run_params.run['n_epochs']):  # loop over the dataset multiple times
+# start the actual training
+net = nn_custom.train_model(net, trainloader, validationloader, scheduler, optimizer,
+                       loss_fn, device, run_params.run['n_epochs'],
+                       run_params.run['plot_every_n_batches'], run_params.run['save_model_every_n_epoch'],
+                       run_params.run['save_params_hist_every_n_epoch'], run_params.run['minibatch_epoch_loss'],
+                       run_params.run['compute_validation_loss'], model_dir, args.use_writer)
 
-    train_loss = 0
-    running_loss = 0.0
-
-    # adjust the learning rate if necessary
-    scheduler.step()
-
-    for i, data in enumerate(trainloader, 0):
-        # get the inputs
-        inputs, labels = data
-        del data
-
-        inputs, labels = inputs.to(device), labels.to(device)
-
-        # zero the parameter gradients
-        optimizer.zero_grad()
-
-        # forward + backward + optimize
-        outputs = net(inputs)
-        del inputs
-        loss = loss_fn(outputs, labels)
-        loss.backward()
-        optimizer.step()
-
-        # print statistics
-        running_loss += loss.item()
-        train_loss += loss.item()
-        del loss
-
-        # print every plot_every_n_batches mini-batches
-        if i % run_params.run['plot_every_n_batches'] == (run_params.run['plot_every_n_batches'] - 1):
-            print('[%d, %5d] averaged loss: %.5f' %
-                  (epoch + 1, i + 1, running_loss / (run_params.run['plot_every_n_batches'])))
-            running_loss = 0.0
-
-    # save model every save_model_every_n_epoch epochs
-    if epoch % run_params.run['save_model_every_n_epoch'] == (run_params.run['save_model_every_n_epoch'] - 1):
-        torch.save(net.state_dict(), os.path.join(model_dir, 'e{}.model'.format(epoch+1)))
-
-    with torch.no_grad():
-        if run_params.run['minibatch_epoch_loss']:
-            train_loss /= len(trainloader)
-        else:
-            train_loss = 0.0
-            for data in trainloader:
-                inputs, labels = data
-                inputs, labels = inputs.to(device), labels.to(device)
-                outputs = net(inputs)
-                loss = loss_fn_val(outputs, labels)
-                train_loss += loss.item()
-            train_loss /= len(trainloader)
-            del data, inputs, labels, outputs
-
-        validation_loss = 0.0
-        if run_params.run['compute_validation_loss']:
-            for data in validationloader:
-                inputs, labels = data
-                inputs, labels = inputs.to(device), labels.to(device)
-                outputs = net(inputs)
-                loss = loss_fn_val(outputs, labels)
-                validation_loss += loss.item()
-            validation_loss /= len(validationloader)
-            del data, inputs, labels, outputs
-
-        writer.add_scalar('Train/Loss', train_loss, epoch+1)
-        writer.add_scalar('Summary/TrainLoss', train_loss, epoch+1)
-        writer.add_scalar('Val/Loss', validation_loss, epoch+1)
-        writer.add_scalar('Summary/ValidationLoss', validation_loss, epoch+1)
-        writer.add_scalar('Summary/LearningRate', scheduler.get_lr()[0], epoch+1)
-
-        if epoch % run_params.run['save_params_hist_every_n_epoch'] == (run_params.run['save_params_hist_every_n_epoch'] - 1):
-            for tag, value in net.named_parameters():
-                tag = tag.replace('.', '/')
-                writer.add_histogram(tag, value.data.cpu().numpy(), epoch+1)
-                writer.add_histogram(tag+'/grad', value.grad.data.cpu().numpy(), epoch+1)
-            del tag, value
-
-        print(('[%d] train loss: %.5f, validation loss: %.5f' %
-                      (epoch + 1, train_loss, validation_loss)))
-
-    gc.collect()
-
-print("INFO: Finished training in %s seconds" % (time.time() - start_time))
-
+# save the model if requested
 if (run_params.run['save_model']):
-    # save the model
     torch.save(net.state_dict(), os.path.join(model_dir, 'latest.model'))
-
-#writer.add_graph(net.cpu(), inputs.cpu()) # do not save the graph by default as the visualization does not work anyway that well
-writer.close()
 
 # evaluate the model performance on the testset if requested
 if (run_params.run['evaluate_testset']):
@@ -229,4 +141,5 @@ if (run_params.run['evaluate_testset']):
 if (os.path.isdir("/cluster/scratch/")):
     os.system('rm '  + trainset_name)
     os.system('rm '  + validationset_name)
-    os.system('rm '  + testset_name)
+    if run_params.run['evaluate_testset']:
+        os.system('rm '  + testset_name)
